+++ src/main/java/com/packsendme/microservice/sa/businessrule/component/BRE_Produce.java
@@ -1,4 +1,4 @@
-package com.packsendme.microservice.sa.component;
+package com.packsendme.microservice.sa.businessrule.component;
 
 public interface BRE_Produce{
