+++ src/main/java/com/packsendme/microservice/sa/businessrule/component/MaritimewayBREConsumer_Component.java
@@ -1,18 +1,18 @@
-package com.packsendme.microservice.sa.component;
+	package com.packsendme.microservice.sa.businessrule.component;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ExecutionBREConsumer_Component implements BRE_Consumer {
+public class MaritimewayBREConsumer_Component implements BRE_Consumer {
 
 	private String msg;
 	
-	@KafkaListener(topics = "${kafka.topic.executionBRE}")
+	@KafkaListener(topics = "${kafka.topic.maritimewayBRE}")
 	public void receive(String data) {
 		this.msg = data;
 		System.out.println(" ------------------------------- ");
-		System.out.println(" topic_executionBRE_sa "+ data);
+		System.out.println(" topic_maritimeway_sa "+ data);
 		System.out.println(" ------------------------------- ");
 	}
