+++ src/main/java/com/packsendme/microservice/sa/businessrule/controller/RoadwayBRE_Controller.java
@@ -1,7 +1,6 @@
 package com.packsendme.microservice.sa.businessrule.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
